Add process- functions to deal with processing parameters -
 When infile= and outstem= were added, the related commands in the loop for
processing the parameters were shifted to the right, making the lines
excessively long.  Put them in functions, at the cost of not being able
to immediately see what hte validation does.

 Also change how the first field of the afade= value is processed, because
a valid value of 'l' (lowercase L) was rejected.

 And if EXPECTEDFRAMES is not set when the input= file is validated, use
its number of frames so that a later validity test will not skip with an
error message from bash.